Added a test for built-in functions
Altered the clock function to return a nicer looking value.

+++ src/com/nervestaple/jlox/interpreter/Interpreter.java
@@ -29,7 +29,7 @@ public class Interpreter implements Expr.Visitor<Object>, Stmt.Visitor<Void> {
 
             @Override
             public Object call(Interpreter interpreter, List<Object> arguments) {
-                return (double) System.currentTimeMillis() / 1000;
+                return System.currentTimeMillis() / 1000;
             }
 
             @Override
